+++ heiya/__init__.py
@@ -1,13 +1,13 @@
 from . import to_hei
 from . import from_hei
-from . import he_jpg
+from . import he_img
 from . import tools
 from . import extensions
 
 __all__ = [
     'to_hei',
     'from_hei',
-    'he_jpg',
+    'he_img',
     'tools',
     'extensions'
 ]